create and use generateClientCommand
